[Refactor] Refactor create protections group
